# perf: use bincode instead of python for io_conf serialization in FileArray