--- a/src/daft-core/src/array/file_array.rs
+++ b/src/daft-core/src/array/file_array.rs
@@ -1,12 +1,8 @@
-#[cfg(feature = "python")]
 use common_file::DaftFileType;
 use common_io_config::IOConfig;
-#[cfg(feature = "python")]
 use daft_schema::{dtype::DataType, field::Field};
 
-#[cfg(feature = "python")]
-use crate::series::IntoSeries;
-use crate::{array::prelude::*, datatypes::FileType};
+use crate::{array::prelude::*, datatypes::FileType, series::IntoSeries};
 
 /// FileArray is a logical array that represents a collection of files.
 ///
@@ -26,14 +22,11 @@ use crate::{array::prelude::*, datatypes::FileType};
 pub type FileArray = LogicalArray<FileType>;
 
 impl FileArray {
-    #[cfg(feature = "python")]
     pub fn new_from_reference_array(
         name: &str,
         urls: &Utf8Array,
         io_config: Option<IOConfig>,
     ) -> Self {
-        use crate::series::IntoSeries;
-
         let discriminant = UInt8Array::from_values(
             "discriminant",
             std::iter::repeat_n(DaftFileType::Reference as u8, urls.len()),
@@ -41,26 +34,12 @@ impl FileArray {
         .into_series();
 
         let sa_field = Field::new("literal", DataType::File.to_physical());
-
-        let io_conf = io_config.map(|c| {
-            let io_conf = common_io_config::python::IOConfig::from(c);
-
-            use std::sync::Arc;
-
-            use pyo3::{IntoPyObjectExt, Python};
-
-            Arc::new(Python::with_gil(|py| {
-                io_conf
-                    .into_py_any(py)
-                    .expect("Failed to convert ioconfig to PyObject")
-            }))
-        });
-
-        let io_configs =
-            PythonArray::from_iter("io_config", std::iter::repeat_n(io_conf, urls.len()));
+        let io_conf: Option<Vec<u8>> =
+            io_config.map(|c| bincode::serialize(&c).expect("Failed to serialize IOConfig"));
+        let io_conf = BinaryArray::from_iter("io_config", std::iter::repeat_n(io_conf, urls.len()));
 
         let data = BinaryArray::full_null("data", &DataType::Binary, urls.len()).into_series();
-        let io_configs = io_configs
+        let io_conf = io_conf
             .with_validity(urls.validity().cloned())
             .expect("Failed to set validity");
 
@@ -70,23 +49,13 @@ impl FileArray {
                 discriminant,
                 data,
                 urls.clone().into_series().rename("url"),
-                io_configs.into_series(),
+                io_conf.into_series(),
             ],
             urls.validity().cloned(),
         );
         FileArray::new(Field::new(name, DataType::File), sa)
     }
 
-    #[cfg(not(feature = "python"))]
-    pub fn new_from_reference_array(
-        name: &str,
-        urls: &Utf8Array,
-        io_config: Option<IOConfig>,
-    ) -> Self {
-        unimplemented!()
-    }
-
-    #[cfg(feature = "python")]
     pub fn new_from_data_array(name: &str, values: &BinaryArray) -> Self {
         let discriminant = UInt8Array::from_values(
             "discriminant",
@@ -97,7 +66,7 @@ impl FileArray {
         let fld = Field::new("literal", DataType::File.to_physical());
         let urls = Utf8Array::full_null("url", &DataType::Utf8, values.len()).into_series();
         let io_configs =
-            PythonArray::full_null("io_config", &DataType::Python, values.len()).into_series();
+            BinaryArray::full_null("io_config", &DataType::Binary, values.len()).into_series();
         let sa = StructArray::new(
             fld,
             vec![
@@ -109,10 +78,5 @@ impl FileArray {
             values.validity().cloned(),
         );
         FileArray::new(Field::new(name, DataType::File), sa)
-    }
-
-    #[cfg(not(feature = "python"))]
-    pub fn new_from_data_array(name: &str, values: &BinaryArray) -> Self {
-        unimplemented!()
     }
 }