## Changes Made

changes out the file array serialization from python to bincode so we don't need to acquire gil. 

This results in about 4-5x faster roundtripping through arrow for the file type. 

and a better cpu utilization. 

~87% on main vs ~99% on feature branch



simple test I used to determine performance.

```py
files = [daft.File("Daft/README.rst", io_config=IOConfig())] * 1_000_000

# create a series of 1 million daft.File's
s = daft.Series.from_pylist(files)
# roundtrip through arrow
s = daft.Series.from_arrow(s.to_arrow())
```



### main: 
```
time python bench.py
python bench.py  9.78s user 0.66s system 87% cpu 11.921 total
```

### feature branch: 
```
time python bench.py
python bench.py  2.36s user 0.30s system 99% cpu 2.673 total
```



## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to `docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly